refactor(apps/frontend-manage): move suspended components to separate files from first login modal
